# src/api/HTEC.POC.API/Controllers/Item/UpdateWalletItemController.cs
using System;
using System.ComponentModel.DataAnnotations;
using System.Threading.Tasks;
using Amido.Stacks.Application.CQRS.Commands;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using HTEC.POC.API.Models.Requests;
using HTEC.POC.CQRS.Commands;

namespace HTEC.POC.API.Controllers;

/// <summary>
/// Item related operations
/// </summary>
[Consumes("application/json")]
[Produces("application/json")]
[ApiExplorerSettings(GroupName = "Item")]
public class UpdateWalletItemController : ApiControllerBase
{
    readonly ICommandHandler<UpdateWalletItem, bool> commandHandler;

    public UpdateWalletItemController(ICommandHandler<UpdateWalletItem, bool> commandHandler)
    {
        this.commandHandler = commandHandler ?? throw new ArgumentNullException(nameof(commandHandler));
    }

    /// <summary>
    /// Update an item in the wallet
    /// </summary>
    /// <remarks>Update an  item in the wallet</remarks>
    /// <param name="id">Id for wallet</param>
    /// <param name="categoryId">Id for Category</param>
    /// <param name="itemId">Id for item being updated</param>
    /// <param name="body">Category being added</param>
    /// <response code="204">No Content</response>
    /// <response code="400">Bad Request</response>
    /// <response code="404">Resource not found</response>
    [HttpPut("/v1/wallet/{id}/category/{categoryId}/items/{itemId}")]
    [Authorize]
    public async Task<IActionResult> UpdateWalletItem([FromRoute][Required]Guid id, [FromRoute][Required]Guid categoryId, [FromRoute][Required]Guid itemId, [FromBody]UpdateItemRequest body)
    {
        // NOTE: Please ensure the API returns the response codes annotated above

        await commandHandler.HandleAsync(
            new UpdateWalletItem(
                correlationId: GetCorrelationId(),
                walletId: id,
                categoryId: categoryId,
                walletItemId: itemId,
                name: body.Name,
                description: body.Description,
                price: body.Price,
                available: body.Available
            )
        );

        return StatusCode(204);
    }
}